UV coords
Models now include UV coordinates. For now, these are stored as integers
and properly scaled in the shader.

+++ src/mdl2gl.hpp
@@ -27,7 +27,7 @@
 struct VertexDef
 {
     GLfloat x, y, z;    // Position
-    GLfloat s, t;       // UV
+    GLint s, t;         // UV
     GLfloat r, g, b;    // Vertex color
 };
